Update dependencies
I had pull requests out for the following packages:

 - teamcity-service-messages
 - markdown-link-check

The first was to include inspection and inspectionType in
the set of available messages. That is now released.

The second was to fix an issue with spaces in baseUrl in
link-check (a dependency of markdown-link-check). That is now
released.

So I can now use the lastest versions of these dependencies.