Allow changed switch to specify base branch
Previously the changed switch (if present) would indicate
to compare current branch against `master` to determine
what files have changed and only check those files.

Now the changed switch can take a branch name:

 --changed=feature1

which means that the base branch for this check should be
`feature1`.

If the changed switch is used, but no branch is specified,
the base branch defaults to `master`.